move forwarding code to own file
